Added exploration logic to RL Agent

+++ ai-agent/train-model.py
@@ -14,9 +14,10 @@ import os
 api_base_url = "http://127.0.0.1:5000"
 out_dir = "./data/training"
 session_template = "training-{:06d}-{:06d}"
-max_epochs = 1000
-max_sessions = 5
+max_epochs = 100
+max_sessions = 10
 
+delete_training_files = True
 # player_server_script = "player-server.py"
 player_server_script = "no-server.py"
 model_filename = "./data/model/t3-simple1.pt"
@@ -28,7 +29,7 @@ def wait_for_server(base_url):
         try:
             response = requests.get(url)
             body = response.json()
-            print(f'Health API response: {body}')
+            # print(f'Health API response: {body}')
 
             if body["alive"]:
                 return
@@ -50,7 +51,7 @@ def reload_model(base_url):
     print(f"Reload API Response: {body}")
 
 
-def run_games(epoch, out_dir):
+def run_games(epoch, out_dir, exploration_rate):
     for i in range(max_sessions):
         session = session_template.format(epoch, i)
         subprocess.run(["node", "../game/build/tic-tac-toe.console.js",
@@ -62,17 +63,20 @@ def run_games(epoch, out_dir):
                         "--configDir", "../game/config",
                         "--outdir", out_dir,
                         "--sessionName", session,
-                        "--encoder", "BitEncoder"
+                        "--encoder", "BitEncoder",
+                        "--explorationRate", f"{exploration_rate}"
                         ])
 
 
-def calculate_reward(action, is_winner, turns_left):
+def calculate_reward(action, player, winner, turns_left):
     if not action["isValid"]:
         return -10
 
-    if is_winner and turns_left == 1:
+    print(f"winner: {winner}, player: {player}, is_winner: {winner == player}, turns_left: {turns_left}")
+
+    if player == winner and turns_left == 1:
         return 1
-    elif not is_winner and turns_left == 2:
+    elif player != winner and turns_left == 2:
         return -1
 
     return 0
@@ -102,9 +106,8 @@ def create_memories(memories, epoch, in_dir):
                 if "choice" not in action:
                     continue
 
-                is_winner = action["player"] == winner
                 turns_left = max_turns - curr_turn
-                reward = calculate_reward(action, is_winner, turns_left)
+                reward = calculate_reward(action, action["player"], winner, turns_left)
 
                 state = [action["player"], action["board"]]
                 choice = action["choice"]
@@ -131,15 +134,18 @@ def make_qlearning_train_step(policy_dqn, target_dqn, loss_fn, optimizer, discou
     def train_step(input, label, reward, next_input):
 
         # Sets model to TRAIN mode
-        policy_dqn.train()
         target_dqn.eval()
+        policy_dqn.train()
 
         # Calculate Q Value
         if next_input is None:
             q_value = torch.tensor(reward)
         else:
             with torch.no_grad():
-                q_value = reward + (discount_rate * target_dqn(next_input).max())
+                # next state's reward is subtracted from current state instead of added
+                # because next state is the other player's turn. Therefore, if the other
+                # player made a successful move, then that is bad for the current player
+                q_value = reward - (discount_rate * target_dqn(next_input).max())
 
         # Makes predictions
         y = policy_dqn(input)
@@ -200,17 +206,20 @@ def train(model, step, memories, decoder=None, board_size=0):
 
 
 def cleanup_files(file_dir, pattern):
-    for f in glob.glob(f"{file_dir}/{pattern}"):
-        os.remove(f)
+    if delete_training_files:
+        for f in glob.glob(f"{file_dir}/{pattern}"):
+            os.remove(f)
 
 
 def app():
+    discovery_rate = 1.0
+    decay_rate = 0.9
+
     learn_rate = 0.01
     discount_rate = 0.9
     policy_dqn = t3.get_model(filename=model_filename)
 
     target_dqn = t3.get_model()
-    target_dqn.load_state_dict(policy_dqn.state_dict())
 
     loss_fn = nn.MSELoss()
     optimizer = optim.Adam(policy_dqn.parameters(), lr=learn_rate)
@@ -223,9 +232,13 @@ def app():
     memories = []
 
     for epoch in range(max_epochs):
-        run_games(epoch, out_dir)
+        target_dqn.load_state_dict(policy_dqn.state_dict())
+        run_games(epoch, out_dir, discovery_rate)
+        discovery_rate = discovery_rate * decay_rate
+
         memories *= 0
         create_memories(memories, epoch, out_dir)
+        print(f"memories={memories}")
 
         train(model=policy_dqn, step=train_step, memories=memories, decoder=encoder, board_size=board_size)
         t3.save_model(policy_dqn, filename=model_filename, archive=False)
